Implemented api endpoints to retrieve and send messages
