+++ src/main/java/xyz/iconc/dev/api/shared/resources/MessageResource.java
@@ -3,7 +3,7 @@ package xyz.iconc.dev.api.shared.resources;
 import org.restlet.resource.Delete;
 import org.restlet.resource.Get;
 import org.restlet.resource.Put;
-import xyz.iconc.dev.api.shared.objects.Message;
+import xyz.iconc.dev.server.apiObjects.Message;
 import xyz.iconc.dev.api.shared.objects.User;
 
 public interface MessageResource extends BaseResource {
