+++ src/test/java/xyz/iconc/dev/server/storage/DatabaseManagerTest.java
@@ -4,9 +4,9 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import xyz.iconc.dev.server.DatabaseManager;
-import xyz.iconc.dev.server.networkObjects.Account;
-import xyz.iconc.dev.server.networkObjects.Channel;
-import xyz.iconc.dev.server.networkObjects.Message;
+import xyz.iconc.dev.server.apiObjects.Account;
+import xyz.iconc.dev.server.apiObjects.Channel;
+import xyz.iconc.dev.server.apiObjects.Message;
 
 class DatabaseManagerTest {
